Enhance WordList class for improved numbering support
Significant additions to the `WordList` class in `WordList.Private.cs` introduce new methods for handling numbering definitions in Word documents. Key functionalities include creating numbering definitions, retrieving the next abstract number and numbering instance, and managing various numbering properties (bold, font size, color, italic, underline, strike, double strike).

Refactored and removed unused properties and methods in `WordList.cs`, changing the class definition to `public partial class WordList` to indicate its split across multiple files. These changes enhance the functionality and organization of the `WordList` class for better handling of numbered lists.
