+++ OfficeIMO.Word/WordList.cs
@@ -1,14 +1,9 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Wordprocessing;
-using SixLabors.ImageSharp;
 
 namespace OfficeIMO.Word;
 
-public class WordList : WordElement {
+public partial class WordList : WordElement {
     private readonly WordprocessingDocument _wordprocessingDocument;
     private readonly WordDocument _document;
     // private readonly WordSection _section;
@@ -33,45 +28,6 @@ public class WordList : WordElement {
                 .Any(style => style.StartsWith("Heading", StringComparison.Ordinal));
         }
     }
-
-    //private string NsidId {
-    //    get {
-    //        if (AbstractNum == null) {
-    //            return null;
-    //        }
-
-    //        return AbstractNum.Nsid.Val;
-
-    //    }
-    //    set {
-    //        if (AbstractNum != null) {
-    //            AbstractNum.Nsid.Val = value;
-    //        }
-    //    }
-    //}
-
-    //private string GenerateNsidId() {
-    //    // https://learn.microsoft.com/en-us/dotnet/api/documentformat.openxml.wordprocessing.nsid?view=openxml-2.8.1
-    //    // Specifies a number value specified as a four digit hexadecimal number),
-    //    // whose contents of this decimal number are interpreted based on the context of the parent XML element.
-    //    // for example FFFFFF89 or D9842532
-    //    return Guid.NewGuid().ToString().ToUpper().Substring(0, 8);
-
-    //}
-
-    //private AbstractNum AbstractNum {
-    //    get {
-    //        var numbering = _document._wordprocessingDocument.MainDocumentPart!.NumberingDefinitionsPart!.Numbering;
-    //        var abstractNumList = numbering.ChildElements.OfType<AbstractNum>();
-    //        foreach (AbstractNum abstractNum in abstractNumList) {
-    //            if (abstractNum.AbstractNumberId == _abstractId) {
-    //                return abstractNum;
-    //            }
-    //        }
-
-    //        return null;
-    //    }
-    //}
 
     public List<WordParagraph> ListItems {
         get {
@@ -248,6 +204,131 @@ public class WordList : WordElement {
         }
     }
 
+    public bool Bold {
+        get { return GetNumberingProperty<bool>(props => props.Elements<Bold>().Any(), false); }
+        set {
+            SetNumberingProperty(props => {
+                props.RemoveAllChildren<Bold>();
+                props.RemoveAllChildren<BoldComplexScript>();
+                if (value) {
+                    props.Append(new Bold());
+                    props.Append(new BoldComplexScript());
+                }
+            }, value);
+        }
+    }
+
+    public int? FontSize {
+        get {
+            return GetNumberingProperty<int?>(props => {
+                var fontSize = props.Elements<FontSize>().FirstOrDefault();
+                // Convert from half-points to points
+                return fontSize?.Val != null ? int.Parse(fontSize.Val) / 2 : null;
+            });
+        }
+        set {
+            SetNumberingProperty(props => {
+                props.RemoveAllChildren<FontSize>();
+                props.RemoveAllChildren<FontSizeComplexScript>();
+                if (value.HasValue) {
+                    // Convert from points to half-points
+                    var halfPoints = (value.Value * 2).ToString();
+                    props.Append(new FontSize { Val = halfPoints });
+                    props.Append(new FontSizeComplexScript { Val = halfPoints });
+                }
+            }, value.HasValue);
+        }
+    }
+
+    public SixLabors.ImageSharp.Color? Color {
+        get {
+            if (ColorHex == "") {
+                return null;
+            }
+            return SixLabors.ImageSharp.Color.Parse("#" + ColorHex);
+        }
+        set {
+            if (value != null) {
+                this.ColorHex = value.Value.ToHexColor();
+            } else {
+                this.ColorHex = "";
+            }
+        }
+    }
+
+    public string ColorHex {
+        get {
+            return GetNumberingProperty<string>(props => {
+                var color = props.Elements<DocumentFormat.OpenXml.Wordprocessing.Color>().FirstOrDefault();
+                return color?.Val ?? "";
+            });
+        }
+        set {
+            SetNumberingProperty(props => {
+                props.RemoveAllChildren<DocumentFormat.OpenXml.Wordprocessing.Color>();
+                if (!string.IsNullOrEmpty(value)) {
+                    props.Append(new DocumentFormat.OpenXml.Wordprocessing.Color {
+                        Val = value.Replace("#", "")
+                    });
+                }
+            }, !string.IsNullOrEmpty(value));
+        }
+    }
+
+    public bool Italic {
+        get => GetNumberingProperty(props => props.Elements<Italic>().Any(), false);
+        set => SetNumberingProperty(props => {
+            props.RemoveAllChildren<Italic>();
+            props.RemoveAllChildren<ItalicComplexScript>();
+            if (value) {
+                props.Append(new Italic());
+                props.Append(new ItalicComplexScript());
+            }
+        }, value);
+    }
+
+    public UnderlineValues? Underline {
+        get => GetNumberingProperty<UnderlineValues?>(props =>
+            props.Elements<Underline>().FirstOrDefault()?.Val);
+        set => SetNumberingProperty(props => {
+            props.RemoveAllChildren<Underline>();
+            if (value.HasValue) {
+                props.Append(new Underline { Val = value.Value });
+            }
+        }, value.HasValue);
+    }
+
+    public bool Strike {
+        get => GetNumberingProperty(props => props.Elements<Strike>().Any(), false);
+        set => SetNumberingProperty(props => {
+            props.RemoveAllChildren<Strike>();
+            if (value) {
+                props.Append(new Strike());
+            }
+        }, value);
+    }
+
+    public bool DoubleStrike {
+        get => GetNumberingProperty(props => props.Elements<DoubleStrike>().Any(), false);
+        set => SetNumberingProperty(props => {
+            props.RemoveAllChildren<DoubleStrike>();
+            if (value) {
+                props.Append(new DoubleStrike());
+            }
+        }, value);
+    }
+
+    public string FontName {
+        get => GetNumberingProperty<string>(props =>
+            props.Elements<RunFonts>().FirstOrDefault()?.Ascii);
+        set => SetNumberingProperty(props => {
+            props.RemoveAllChildren<RunFonts>();
+            if (!string.IsNullOrEmpty(value)) {
+                props.Append(new RunFonts { Ascii = value, HighAnsi = value });
+            }
+        }, !string.IsNullOrEmpty(value));
+    }
+
     public WordList(WordDocument wordDocument, bool isToc = false) {
         _document = wordDocument;
         _wordprocessingDocument = wordDocument._wordprocessingDocument;
@@ -362,193 +443,7 @@ public class WordList : WordElement {
 
         return wordParagraph;
     }
-
-    private static int GetNextAbstractNum(Numbering numbering) {
-        var ids = numbering.ChildElements
-            .OfType<AbstractNum>()
-            .Select(element => (int)element.AbstractNumberId)
-            .ToList();
-        return ids.Count > 0 ? ids.Max() + 1 : 0;
-    }
-
-    private static int GetNextNumberingInstance(Numbering numbering) {
-        var ids = numbering.ChildElements
-            .OfType<NumberingInstance>()
-            .Select(element => (int)element.NumberID)
-            .ToList();
-        return ids.Count > 0 ? ids.Max() + 1 : 1;
-    }
-
-    internal void AddList(WordListStyle style) {
-        CreateNumberingDefinition(_document);
-        var numbering = _document._wordprocessingDocument.MainDocumentPart!.NumberingDefinitionsPart!.Numbering;
-
-        _abstractId = GetNextAbstractNum(numbering);
-        _numberId = GetNextNumberingInstance(numbering);
-
-        var abstractNum = WordListStyles.GetStyle(style);
-        abstractNum.AbstractNumberId = _abstractId;
-        var abstractNumId = new AbstractNumId {
-            Val = _abstractId
-        };
-        NumberingInstance numberingInstance = new NumberingInstance();
-        numberingInstance = RestartNumberingInstance(abstractNumId, _numberId);
-        numbering.Append(numberingInstance, abstractNum);
-    }
-
-    // TODO this isn't working yet, needs implementation
-    //internal void AddList(CustomListStyles style = CustomListStyles.Bullet, string levelText = "·", int levelIndex = 0) {
-    //    CreateNumberingDefinition(_document);
-
-    //    // we take current list number from the document
-    //    //_numberId = _document._listNumbers;
-
-    //    var numberingFormatValues = CustomListStyle.GetStyle(style);
-
-    //    var level = new Level(
-    //        new NumberingFormat { Val = numberingFormatValues },
-    //        new LevelText { Val = levelText }
-    //    ) {
-    //        LevelIndex = 1
-    //    };
-    //    var level1 = new Level(
-    //        new NumberingFormat { Val = numberingFormatValues },
-    //        new LevelText { Val = levelText }
-    //    ) {
-    //        LevelIndex = 2
-    //    };
-    //    var abstractNum = new AbstractNum(level, level1) {
-    //        AbstractNumberId = 0
-    //    };
-    //    //abstractNum.Nsid = new Nsid();
-
-    //    var numbering = _document._wordprocessingDocument.MainDocumentPart!.NumberingDefinitionsPart!.Numbering;
-    //    numbering.Append(abstractNum);
-
-    //    var abstractNumId = new AbstractNumId {
-    //        Val = 0
-    //    };
-    //    var numberingInstance = new NumberingInstance(abstractNumId) {
-    //        NumberID = _numberId
-    //    };
-
-    //    //LevelOverride levelOverride = new LevelOverride();
-    //    //levelOverride.StartOverrideNumberingValue = new StartOverrideNumberingValue();
-    //    //levelOverride.StartOverrideNumberingValue.Val = 1;
-    //    //numberingInstance.Append(levelOverride);
-
-    //    numbering.Append(numberingInstance);
-    //}
-
-    private void CreateNumberingDefinition(WordDocument document) {
-        var numberingDefinitionsPart = document._wordprocessingDocument.MainDocumentPart!.NumberingDefinitionsPart ?? _wordprocessingDocument.MainDocumentPart!.AddNewPart<NumberingDefinitionsPart>();
-        if (numberingDefinitionsPart.Numbering == null) {
-            // the check for null is required even tho Resharper claims it's not
-            Numbering numbering1 = new Numbering() { MCAttributes = new MarkupCompatibilityAttributes() { Ignorable = "w14 w15 w16se w16cid w16 w16cex w16sdtdh wp14" } };
-            numbering1.AddNamespaceDeclaration("wpc", "http://schemas.microsoft.com/office/word/2010/wordprocessingCanvas");
-            numbering1.AddNamespaceDeclaration("cx", "http://schemas.microsoft.com/office/drawing/2014/chartex");
-            numbering1.AddNamespaceDeclaration("cx1", "http://schemas.microsoft.com/office/drawing/2015/9/8/chartex");
-            numbering1.AddNamespaceDeclaration("cx2", "http://schemas.microsoft.com/office/drawing/2015/10/21/chartex");
-            numbering1.AddNamespaceDeclaration("cx3", "http://schemas.microsoft.com/office/drawing/2016/5/9/chartex");
-            numbering1.AddNamespaceDeclaration("cx4", "http://schemas.microsoft.com/office/drawing/2016/5/10/chartex");
-            numbering1.AddNamespaceDeclaration("cx5", "http://schemas.microsoft.com/office/drawing/2016/5/11/chartex");
-            numbering1.AddNamespaceDeclaration("cx6", "http://schemas.microsoft.com/office/drawing/2016/5/12/chartex");
-            numbering1.AddNamespaceDeclaration("cx7", "http://schemas.microsoft.com/office/drawing/2016/5/13/chartex");
-            numbering1.AddNamespaceDeclaration("cx8", "http://schemas.microsoft.com/office/drawing/2016/5/14/chartex");
-            numbering1.AddNamespaceDeclaration("mc", "http://schemas.openxmlformats.org/markup-compatibility/2006");
-            numbering1.AddNamespaceDeclaration("aink", "http://schemas.microsoft.com/office/drawing/2016/ink");
-            numbering1.AddNamespaceDeclaration("am3d", "http://schemas.microsoft.com/office/drawing/2017/model3d");
-            numbering1.AddNamespaceDeclaration("o", "urn:schemas-microsoft-com:office:office");
-            numbering1.AddNamespaceDeclaration("oel", "http://schemas.microsoft.com/office/2019/extlst");
-            numbering1.AddNamespaceDeclaration("r", "http://schemas.openxmlformats.org/officeDocument/2006/relationships");
-            numbering1.AddNamespaceDeclaration("m", "http://schemas.openxmlformats.org/officeDocument/2006/math");
-            numbering1.AddNamespaceDeclaration("v", "urn:schemas-microsoft-com:vml");
-            numbering1.AddNamespaceDeclaration("wp14", "http://schemas.microsoft.com/office/word/2010/wordprocessingDrawing");
-            numbering1.AddNamespaceDeclaration("wp", "http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing");
-            numbering1.AddNamespaceDeclaration("w10", "urn:schemas-microsoft-com:office:word");
-            numbering1.AddNamespaceDeclaration("w", "http://schemas.openxmlformats.org/wordprocessingml/2006/main");
-            numbering1.AddNamespaceDeclaration("w14", "http://schemas.microsoft.com/office/word/2010/wordml");
-            numbering1.AddNamespaceDeclaration("w15", "http://schemas.microsoft.com/office/word/2012/wordml");
-            numbering1.AddNamespaceDeclaration("w16cex", "http://schemas.microsoft.com/office/word/2018/wordml/cex");
-            numbering1.AddNamespaceDeclaration("w16cid", "http://schemas.microsoft.com/office/word/2016/wordml/cid");
-            numbering1.AddNamespaceDeclaration("w16", "http://schemas.microsoft.com/office/word/2018/wordml");
-            numbering1.AddNamespaceDeclaration("w16sdtdh", "http://schemas.microsoft.com/office/word/2020/wordml/sdtdatahash");
-            numbering1.AddNamespaceDeclaration("w16se", "http://schemas.microsoft.com/office/word/2015/wordml/symex");
-            numbering1.AddNamespaceDeclaration("wpg", "http://schemas.microsoft.com/office/word/2010/wordprocessingGroup");
-            numbering1.AddNamespaceDeclaration("wpi", "http://schemas.microsoft.com/office/word/2010/wordprocessingInk");
-            numbering1.AddNamespaceDeclaration("wne", "http://schemas.microsoft.com/office/word/2006/wordml");
-            numbering1.AddNamespaceDeclaration("wps", "http://schemas.microsoft.com/office/word/2010/wordprocessingShape");
-
-            numberingDefinitionsPart.Numbering = numbering1;
-            numberingDefinitionsPart.Numbering.Save(_document._wordprocessingDocument.MainDocumentPart.NumberingDefinitionsPart);
-        }
-    }
-
-    private NumberingInstance DefaultNumberingInstance(AbstractNumId abstractNumId, int numberId) {
-        var numberingInstance = new NumberingInstance(abstractNumId) { NumberID = numberId };
-        return numberingInstance;
-    }
-
-    private NumberingInstance RestartNumberingInstance(AbstractNumId abstractNumId, int numberId) {
-        NumberingInstance numberingInstance1 = new NumberingInstance(abstractNumId) { NumberID = numberId };
-
-        LevelOverride levelOverride1 = new LevelOverride() { LevelIndex = 0 };
-        StartOverrideNumberingValue startOverrideNumberingValue1 = new StartOverrideNumberingValue() { Val = 1 };
-
-        levelOverride1.Append(startOverrideNumberingValue1);
-
-        LevelOverride levelOverride2 = new LevelOverride() { LevelIndex = 1 };
-        StartOverrideNumberingValue startOverrideNumberingValue2 = new StartOverrideNumberingValue() { Val = 1 };
-
-        levelOverride2.Append(startOverrideNumberingValue2);
-
-        LevelOverride levelOverride3 = new LevelOverride() { LevelIndex = 2 };
-        StartOverrideNumberingValue startOverrideNumberingValue3 = new StartOverrideNumberingValue() { Val = 1 };
-
-        levelOverride3.Append(startOverrideNumberingValue3);
-
-        LevelOverride levelOverride4 = new LevelOverride() { LevelIndex = 3 };
-        StartOverrideNumberingValue startOverrideNumberingValue4 = new StartOverrideNumberingValue() { Val = 1 };
-
-        levelOverride4.Append(startOverrideNumberingValue4);
-
-        LevelOverride levelOverride5 = new LevelOverride() { LevelIndex = 4 };
-        StartOverrideNumberingValue startOverrideNumberingValue5 = new StartOverrideNumberingValue() { Val = 1 };
-
-        levelOverride5.Append(startOverrideNumberingValue5);
-
-        LevelOverride levelOverride6 = new LevelOverride() { LevelIndex = 5 };
-        StartOverrideNumberingValue startOverrideNumberingValue6 = new StartOverrideNumberingValue() { Val = 1 };
-
-        levelOverride6.Append(startOverrideNumberingValue6);
-
-        LevelOverride levelOverride7 = new LevelOverride() { LevelIndex = 6 };
-        StartOverrideNumberingValue startOverrideNumberingValue7 = new StartOverrideNumberingValue() { Val = 1 };
-
-        levelOverride7.Append(startOverrideNumberingValue7);
-
-        LevelOverride levelOverride8 = new LevelOverride() { LevelIndex = 7 };
-        StartOverrideNumberingValue startOverrideNumberingValue8 = new StartOverrideNumberingValue() { Val = 1 };
-
-        levelOverride8.Append(startOverrideNumberingValue8);
-
-        LevelOverride levelOverride9 = new LevelOverride() { LevelIndex = 8 };
-        StartOverrideNumberingValue startOverrideNumberingValue9 = new StartOverrideNumberingValue() { Val = 1 };
-
-        levelOverride9.Append(startOverrideNumberingValue9);
-
-        numberingInstance1.Append(levelOverride1);
-        numberingInstance1.Append(levelOverride2);
-        numberingInstance1.Append(levelOverride3);
-        numberingInstance1.Append(levelOverride4);
-        numberingInstance1.Append(levelOverride5);
-        numberingInstance1.Append(levelOverride6);
-        numberingInstance1.Append(levelOverride7);
-        numberingInstance1.Append(levelOverride8);
-        numberingInstance1.Append(levelOverride9);
-        return numberingInstance1;
-    }
-
+    
     public void Remove() {
         // Get the Numbering part from the document
         var numbering = _document._wordprocessingDocument.MainDocumentPart.NumberingDefinitionsPart.Numbering;
